test: add small constant epsilon for result approximation
Add a constant value to the approximation threshold for floating-point
numbers.

Clang on AArch64 compiles `vec_cross` with optimizations that cause
certain fields that should be zero, to be off from zero by sometimes as
much as 1.0f. These fields were then compared with an epsilon calculated
by multiplying with the expected field value, yielding exactly zero, and
failing the comparison.

"Fix" this by adding a slight constant offset. The proper fix will be to
add a native AArch64 `vec_cross' function.